tyck: move `TyckState`
