+++ base/src/main/java/org/aya/core/Meta.java
@@ -12,7 +12,7 @@ import org.aya.generic.util.NormalizeMode;
 import org.aya.util.Arg;
 import org.aya.generic.Constants;
 import org.aya.ref.AnyVar;
-import org.aya.tyck.TyckState;
+import org.aya.tyck.tycker.TyckState;
 import org.aya.util.error.SourcePos;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
