+++ base/src/main/java/org/aya/tyck/LittleTyper.java
@@ -13,6 +13,7 @@ import org.aya.generic.util.InternalException;
 import org.aya.generic.util.NormalizeMode;
 import org.aya.guest0x0.cubical.Partial;
 import org.aya.tyck.env.LocalCtx;
+import org.aya.tyck.tycker.TyckState;
 import org.aya.util.Arg;
 import org.jetbrains.annotations.NotNull;
