+++ base/src/main/java/org/aya/tyck/pat/PatClassifier.java
@@ -19,7 +19,7 @@ import org.aya.core.visitor.Subst;
 import org.aya.generic.util.NormalizeMode;
 import org.aya.ref.AnyVar;
 import org.aya.tyck.ExprTycker;
-import org.aya.tyck.TyckState;
+import org.aya.tyck.tycker.TyckState;
 import org.aya.tyck.env.LocalCtx;
 import org.aya.tyck.error.TyckOrderError;
 import org.aya.tyck.error.UnifyInfo;
