+++ base/src/main/java/org/aya/tyck/unify/TermComparator.java
@@ -26,7 +26,7 @@ import org.aya.guest0x0.cubical.Restr;
 import org.aya.ref.AnyVar;
 import org.aya.ref.DefVar;
 import org.aya.ref.LocalVar;
-import org.aya.tyck.TyckState;
+import org.aya.tyck.tycker.TyckState;
 import org.aya.tyck.env.LocalCtx;
 import org.aya.tyck.error.LevelError;
 import org.aya.tyck.trace.Trace;
